Weird time and code fix
Since token should use unix timestamp, using the tick_time for that
doesn't make sense because it is adjusted for timezone.  And since there
is no function that would transfer the localtime to the timestamp, we
should rather call time(NULL) again.

What makes this change really intereseting is that before this change,
the watchface would show wrong time -- UTC, not local.  After this
change, it shows proper time using the timezone that is set and uses
unix timestamp for the token calculation so that it is not related to
timezone, hence users can travel wherever they want.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>